--- src/main/java/de/dhbw/ast/Clazz.java
@@ -1,6 +1,0 @@
-package de.dhbw.ast;
-
-import java.util.List;
-
-public record Clazz(Type name, List<Field> fields) {
-}
