--- src/main/java/de/dhbw/ast/Field.java
@@ -1,4 +1,0 @@
-package de.dhbw.ast;
-
-public record Field(Type type, String name) {
-}
