--- src/main/java/de/dhbw/ast/Parameter.java
@@ -1,4 +1,0 @@
-package de.dhbw.ast;
-
-public record Parameter(Type type, String name) {
-}
