--- src/main/java/de/dhbw/ast/Program.java
@@ -1,6 +1,0 @@
-package de.dhbw.ast;
-
-import java.util.List;
-
-public record Program(List<Clazz> classes) {
-}
