--- src/main/java/de/dhbw/ast/Type.java
@@ -1,4 +1,0 @@
-package de.dhbw.ast;
-
-public record Type() {
-}
