--- src/main/java/de/dhbw/ast/expressions/Binary.java
@@ -1,4 +1,0 @@
-package de.dhbw.ast.expressions;
-
-public record Binary(String operator, Expression left, Expression right) implements Expression {
-}
