--- src/main/java/de/dhbw/ast/expressions/Bool.java
@@ -1,4 +1,0 @@
-package de.dhbw.ast.expressions;
-
-public record Bool(Boolean value) implements Expression {
-}
