--- src/main/java/de/dhbw/ast/expressions/Char.java
@@ -1,4 +1,0 @@
-package de.dhbw.ast.expressions;
-
-public record Char(Character value) implements Expression {
-}
