--- src/main/java/de/dhbw/ast/expressions/Expression.java
@@ -1,5 +1,0 @@
-package de.dhbw.ast.expressions;
-
-public sealed interface Expression permits
-    This, Super, LocalOrFieldVar, InstVar, Unary, Binary, Integr, Bool, Char, Strng, Jnull, StmtExprExpr {
-}
