--- src/main/java/de/dhbw/ast/expressions/InstVar.java
@@ -1,4 +1,0 @@
-package de.dhbw.ast.expressions;
-
-public record InstVar(Expression thisExpr, String varName) implements Expression {
-}
