--- src/main/java/de/dhbw/ast/expressions/Integr.java
@@ -1,4 +1,0 @@
-package de.dhbw.ast.expressions;
-
-public record Integr(Integer value) implements Expression {
-}
