--- src/main/java/de/dhbw/ast/expressions/Jnull.java
@@ -1,4 +1,0 @@
-package de.dhbw.ast.expressions;
-
-public record Jnull() implements Expression {
-}
