--- src/main/java/de/dhbw/ast/expressions/LocalOrFieldVar.java
@@ -1,4 +1,0 @@
-package de.dhbw.ast.expressions;
-
-public record LocalOrFieldVar(String name) implements Expression {
-}
