--- src/main/java/de/dhbw/ast/expressions/StmtExprExpr.java
@@ -1,6 +1,0 @@
-package de.dhbw.ast.expressions;
-
-import de.dhbw.ast.stmtexprs.StatementExpression;
-
-public record StmtExprExpr(StatementExpression statementExpression) implements Expression {
-}
