--- src/main/java/de/dhbw/ast/expressions/Strng.java
@@ -1,4 +1,0 @@
-package de.dhbw.ast.expressions;
-
-public record Strng(String value) implements Expression {
-}
