--- src/main/java/de/dhbw/ast/expressions/Super.java
@@ -1,4 +1,0 @@
-package de.dhbw.ast.expressions;
-
-public record Super() implements Expression {
-}
