--- src/main/java/de/dhbw/ast/expressions/This.java
@@ -1,4 +1,0 @@
-package de.dhbw.ast.expressions;
-
-public record This() implements Expression {
-}
