--- src/main/java/de/dhbw/ast/expressions/Unary.java
@@ -1,4 +1,0 @@
-package de.dhbw.ast.expressions;
-
-public record Unary(String operator, Expression argument) implements Expression {
-}
