--- src/main/java/de/dhbw/ast/statements/Block.java
@@ -1,6 +1,0 @@
-package de.dhbw.ast.statements;
-
-import java.util.List;
-
-public record Block(List<Statement> stmts) implements Statement {
-}
