--- src/main/java/de/dhbw/ast/statements/If.java
@@ -1,6 +1,0 @@
-package de.dhbw.ast.statements;
-
-import de.dhbw.ast.expressions.Expression;
-
-public record If(Expression condition, Statement ifBody, Statement elseBody) implements Statement {
-}
