--- src/main/java/de/dhbw/ast/statements/LocalVarDecl.java
@@ -1,6 +1,0 @@
-package de.dhbw.ast.statements;
-
-import de.dhbw.ast.Type;
-
-public record LocalVarDecl(Type type, String name) implements Statement {
-}
