--- src/main/java/de/dhbw/ast/statements/Return.java
@@ -1,6 +1,0 @@
-package de.dhbw.ast.statements;
-
-import de.dhbw.ast.expressions.Expression;
-
-public record Return(Expression expression) implements Statement {
-}
