--- src/main/java/de/dhbw/ast/statements/Statement.java
@@ -1,4 +1,0 @@
-package de.dhbw.ast.statements;
-
-public sealed interface Statement permits Block, Return, While, LocalVarDecl, If, StmtExprStmt {
-}
