--- src/main/java/de/dhbw/ast/statements/StmtExprStmt.java
@@ -1,6 +1,0 @@
-package de.dhbw.ast.statements;
-
-import de.dhbw.ast.stmtexprs.StatementExpression;
-
-public record StmtExprStmt(StatementExpression statementExpression) implements Statement {
-}
