--- src/main/java/de/dhbw/ast/statements/While.java
@@ -1,6 +1,0 @@
-package de.dhbw.ast.statements;
-
-import de.dhbw.ast.expressions.Expression;
-
-public record While(Expression condition, Statement statement) implements Statement {
-}
