--- src/main/java/de/dhbw/ast/stmtexprs/Assign.java
@@ -1,6 +1,0 @@
-package de.dhbw.ast.stmtexprs;
-
-import de.dhbw.ast.expressions.Expression;
-
-public record Assign(String var, Expression value) implements StatementExpression {
-}
