--- src/main/java/de/dhbw/ast/stmtexprs/MethodCall.java
@@ -1,8 +1,0 @@
-package de.dhbw.ast.stmtexprs;
-
-import de.dhbw.ast.expressions.Expression;
-
-import java.util.List;
-
-public record MethodCall(Expression thisExpr, String name, List<Expression> args) implements StatementExpression {
-}
