--- src/main/java/de/dhbw/ast/stmtexprs/New.java
@@ -1,7 +1,0 @@
-package de.dhbw.ast.stmtexprs;
-
-import de.dhbw.ast.expressions.Expression;
-import de.dhbw.ast.Type;
-
-public record New(Type type, Expression expression) implements StatementExpression {
-}
