--- src/main/java/de/dhbw/ast/stmtexprs/StatementExpression.java
@@ -1,4 +1,0 @@
-package de.dhbw.ast.stmtexprs;
-
-public sealed interface StatementExpression permits Assign, New, MethodCall {
-}
