+++ src/main/java/de/dhbw/compiler/ast/Clazz.java
@@ -1,0 +1,6 @@
+package de.dhbw.compiler.ast;
+
+import java.util.List;
+
+public record Clazz(Type name, List<Field> fields) {
+}
