+++ src/main/java/de/dhbw/compiler/ast/Field.java
@@ -1,0 +1,4 @@
+package de.dhbw.compiler.ast;
+
+public record Field(Type type, String name) {
+}
