+++ src/main/java/de/dhbw/compiler/ast/Method.java
@@ -1,0 +1,8 @@
+package de.dhbw.compiler.ast;
+
+import de.dhbw.compiler.ast.statements.Statement;
+
+import java.util.List;
+
+public record Method(Type returnType, String name, List<Parameter> parameters, Statement statement) {
+}
