+++ src/main/java/de/dhbw/compiler/ast/Parameter.java
@@ -1,0 +1,4 @@
+package de.dhbw.compiler.ast;
+
+public record Parameter(Type type, String name) {
+}
