+++ src/main/java/de/dhbw/compiler/ast/Program.java
@@ -1,0 +1,6 @@
+package de.dhbw.compiler.ast;
+
+import java.util.List;
+
+public record Program(List<Clazz> classes) {
+}
