+++ src/main/java/de/dhbw/compiler/ast/Type.java
@@ -1,0 +1,4 @@
+package de.dhbw.compiler.ast;
+
+public record Type() {
+}
