+++ src/main/java/de/dhbw/compiler/ast/expressions/Binary.java
@@ -1,0 +1,4 @@
+package de.dhbw.compiler.ast.expressions;
+
+public record Binary(String operator, Expression left, Expression right) implements Expression {
+}
