+++ src/main/java/de/dhbw/compiler/ast/expressions/Bool.java
@@ -1,0 +1,4 @@
+package de.dhbw.compiler.ast.expressions;
+
+public record Bool(Boolean value) implements Expression {
+}
