+++ src/main/java/de/dhbw/compiler/ast/expressions/Char.java
@@ -1,0 +1,4 @@
+package de.dhbw.compiler.ast.expressions;
+
+public record Char(Character value) implements Expression {
+}
