+++ src/main/java/de/dhbw/compiler/ast/expressions/Expression.java
@@ -1,0 +1,5 @@
+package de.dhbw.compiler.ast.expressions;
+
+public sealed interface Expression permits
+    This, Super, LocalOrFieldVar, InstVar, Unary, Binary, Integr, Bool, Char, Strng, Jnull, StmtExprExpr {
+}
