+++ src/main/java/de/dhbw/compiler/ast/expressions/InstVar.java
@@ -1,0 +1,4 @@
+package de.dhbw.compiler.ast.expressions;
+
+public record InstVar(Expression thisExpr, String varName) implements Expression {
+}
