+++ src/main/java/de/dhbw/compiler/ast/expressions/Integr.java
@@ -1,0 +1,4 @@
+package de.dhbw.compiler.ast.expressions;
+
+public record Integr(Integer value) implements Expression {
+}
