+++ src/main/java/de/dhbw/compiler/ast/expressions/LocalOrFieldVar.java
@@ -1,0 +1,4 @@
+package de.dhbw.compiler.ast.expressions;
+
+public record LocalOrFieldVar(String name) implements Expression {
+}
