+++ src/main/java/de/dhbw/compiler/ast/expressions/StmtExprExpr.java
@@ -1,0 +1,6 @@
+package de.dhbw.compiler.ast.expressions;
+
+import de.dhbw.compiler.ast.stmtexprs.StatementExpression;
+
+public record StmtExprExpr(StatementExpression statementExpression) implements Expression {
+}
