+++ src/main/java/de/dhbw/compiler/ast/expressions/Strng.java
@@ -1,0 +1,4 @@
+package de.dhbw.compiler.ast.expressions;
+
+public record Strng(String value) implements Expression {
+}
