+++ src/main/java/de/dhbw/compiler/ast/expressions/This.java
@@ -1,0 +1,4 @@
+package de.dhbw.compiler.ast.expressions;
+
+public record This() implements Expression {
+}
