+++ src/main/java/de/dhbw/compiler/ast/expressions/Unary.java
@@ -1,0 +1,4 @@
+package de.dhbw.compiler.ast.expressions;
+
+public record Unary(String operator, Expression argument) implements Expression {
+}
