+++ src/main/java/de/dhbw/compiler/ast/statements/Block.java
@@ -1,0 +1,6 @@
+package de.dhbw.compiler.ast.statements;
+
+import java.util.List;
+
+public record Block(List<Statement> stmts) implements Statement {
+}
