+++ src/main/java/de/dhbw/compiler/ast/statements/LocalVarDecl.java
@@ -1,0 +1,6 @@
+package de.dhbw.compiler.ast.statements;
+
+import de.dhbw.compiler.ast.Type;
+
+public record LocalVarDecl(Type type, String name) implements Statement {
+}
