+++ src/main/java/de/dhbw/compiler/ast/statements/Return.java
@@ -1,0 +1,6 @@
+package de.dhbw.compiler.ast.statements;
+
+import de.dhbw.compiler.ast.expressions.Expression;
+
+public record Return(Expression expression) implements Statement {
+}
