+++ src/main/java/de/dhbw/compiler/ast/statements/Statement.java
@@ -1,0 +1,4 @@
+package de.dhbw.compiler.ast.statements;
+
+public sealed interface Statement permits Block, Return, While, LocalVarDecl, If, StmtExprStmt {
+}
