+++ src/main/java/de/dhbw/compiler/ast/statements/StmtExprStmt.java
@@ -1,0 +1,6 @@
+package de.dhbw.compiler.ast.statements;
+
+import de.dhbw.compiler.ast.stmtexprs.StatementExpression;
+
+public record StmtExprStmt(StatementExpression statementExpression) implements Statement {
+}
