+++ src/main/java/de/dhbw/compiler/ast/statements/While.java
@@ -1,0 +1,6 @@
+package de.dhbw.compiler.ast.statements;
+
+import de.dhbw.compiler.ast.expressions.Expression;
+
+public record While(Expression condition, Statement statement) implements Statement {
+}
