+++ src/main/java/de/dhbw/compiler/ast/stmtexprs/Assign.java
@@ -1,0 +1,6 @@
+package de.dhbw.compiler.ast.stmtexprs;
+
+import de.dhbw.compiler.ast.expressions.Expression;
+
+public record Assign(String var, Expression value) implements StatementExpression {
+}
