+++ src/main/java/de/dhbw/compiler/ast/stmtexprs/New.java
@@ -1,0 +1,7 @@
+package de.dhbw.compiler.ast.stmtexprs;
+
+import de.dhbw.compiler.ast.expressions.Expression;
+import de.dhbw.compiler.ast.Type;
+
+public record New(Type type, Expression expression) implements StatementExpression {
+}
