+++ src/main/java/de/dhbw/compiler/ast/stmtexprs/StatementExpression.java
@@ -1,0 +1,4 @@
+package de.dhbw.compiler.ast.stmtexprs;
+
+public sealed interface StatementExpression permits Assign, New, MethodCall {
+}
